Controller Artist + templates READ

+++ src/main/java/projet/pid/reservationsspringboot/repository/ArtistRepository.java
@@ -4,8 +4,9 @@ import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 import projet.pid.reservationsspringboot.model.Artist;
 
-public interface ArtistRepository extends CrudRepository<Artist, Long> {
-    
+public interface ArtistRepository extends CrudRepository<Artist, Integer> {
+
     List<Artist> findByLastname(String lastname);
-    Artist findById(long id);
+
+    Artist findById(int id);
 }
